% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.R
\name{KL_for_sales}
\alias{KL_for_sales}
\title{KL function for calculating the KL loss for weighted mixture models.}
\usage{
KL_for_sales(i, train, test, K = 7, R = 2)
}
\arguments{
\item{i}{The current model index.}

\item{train}{The training dataset.}

\item{test}{The test dataset.}

\item{K}{The number of candidate models to fit (default is 7).}

\item{R}{The range for model selection based on AIC/BIC (default is 2).}
}
\value{
A list containing the adjusted KL losses, individual model KL losses, AIC and BIC weights, and CV weights for both 5 and 10 folds \code{n}
}
\description{
This function computes the KL divergence for each model, including AIC, BIC, and weighted models, and performs cross-validation. 
It returns adjusted KL losses, weights, and model selection results based on AIC, BIC, and cross-validation.
}
